+++ test_example/test_example/asgi.py
@@ -1,3 +1,7 @@
+import os
+
 from django.core.asgi import get_asgi_application
+
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "test_example.settings")
 
 application = get_asgi_application()
