+++ test_example/test_example/settings.py
@@ -1,0 +1,35 @@
+import os
+from pathlib import Path
+
+BASE_DIR = Path(__file__).resolve().parent.parent
+SECRET_KEY = os.environ.get("DJANGO_SECRET_KEY", "unsafe-default-key")
+DEBUG = True
+
+ALLOWED_HOSTS = ["*"]
+
+INSTALLED_APPS = [
+    "books",
+]
+
+
+ROOT_URLCONF = "test_example.urls"
+
+DATABASES = {
+    "default": {
+        "ENGINE": "django_async_backend.db.backends.postgresql",
+        "NAME": "postgres",
+        "USER": "postgres",
+        "PASSWORD": "postgres",
+        "HOST": "localhost",
+        "PORT": 5432,
+        "OPTIONS": {
+            "pool": {
+                "min_size": 2,
+                "max_size": 4,
+                "timeout": 60,
+            }
+        },
+    }
+}
+
+WSGI_APPLICATION = "test_example.wsgi.application"
