+++ test_example/test_example/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from test_example.views import index
+
+urlpatterns = [
+    path("", index),
+]
